+++ RandomSkunk.Results.UnitTests/SelectMany_Async_methods.cs
@@ -132,7 +132,7 @@ public class SelectMany_Async_methods
         [Fact]
         public async Task Given_target_is_Maybe_of_T_When_source_is_None_Returns_None()
         {
-            var source = Maybe<int>.None();
+            var source = Maybe<int>.None;
 
             var actual = await source.SelectMany(value => Task.FromResult(value.ToString().ToMaybe()));
 
@@ -163,7 +163,7 @@ public class SelectMany_Async_methods
         [Fact]
         public async Task Given_target_is_Result_When_source_is_None_Returns_Fail_result()
         {
-            var source = Maybe<string>.None();
+            var source = Maybe<string>.None;
 
             var actual = await source.SelectMany(value => Task.FromResult(Result.Success()));
 
@@ -212,7 +212,7 @@ public class SelectMany_Async_methods
         [Fact]
         public async Task Given_target_is_Result_of_T_When_source_is_None_Returns_Fail_result()
         {
-            var source = Maybe<int>.None();
+            var source = Maybe<int>.None;
 
             var actual = await source.SelectMany(value => Task.FromResult(value.ToString().ToResult()));