+++ RandomSkunk.Results/Linq/EnumerableExtensions.cs
@@ -214,7 +214,7 @@ public static class EnumerableExtensions
                 return enumerator.Current.ToMaybeOrFailIfNull("The first element was null.");
         }
 
-        return Maybe<T>.None();
+        return Maybe<T>.None;
     }
 
     /// <summary>
@@ -265,7 +265,7 @@ public static class EnumerableExtensions
                 return item.ToMaybeOrFailIfNull("The first matching element was null.");
         }
 
-        return Maybe<T>.None();
+        return Maybe<T>.None;
     }
 
     /// <summary>
@@ -337,7 +337,7 @@ public static class EnumerableExtensions
             }
         }
 
-        return Maybe<T>.None();
+        return Maybe<T>.None;
     }
 
     /// <summary>
@@ -380,7 +380,7 @@ public static class EnumerableExtensions
         if (sourceSequence is null) throw new ArgumentNullException(nameof(sourceSequence));
         if (predicate is null) throw new ArgumentNullException(nameof(predicate));
 
-        var result = Maybe<T>.None();
+        var result = Maybe<T>.None;
         foreach (T item in sourceSequence)
             if (predicate(item)) result = item.ToMaybeOrFailIfNull("The last matching element was null.");
 
@@ -445,7 +445,7 @@ public static class EnumerableExtensions
             var count = list.Count;
             return count switch
             {
-                0 => Maybe<T>.None(),
+                0 => Maybe<T>.None,
                 1 => list[0].ToMaybeOrFailIfNull("The single element was null."),
                 _ => Maybe<T>.Fail(SequenceContainsMoreThanOneElement(count)),
             };
@@ -454,7 +454,7 @@ public static class EnumerableExtensions
         {
             using IEnumerator<T> enumerator = sourceSequence.GetEnumerator();
             if (!enumerator.MoveNext())
-                return Maybe<T>.None();
+                return Maybe<T>.None;
 
             T current = enumerator.Current;
             if (!enumerator.MoveNext())
@@ -532,7 +532,7 @@ public static class EnumerableExtensions
 
         return count switch
         {
-            0 => Maybe<T>.None(),
+            0 => Maybe<T>.None,
             1 => result,
             _ => Maybe<T>.Fail(SequenceContainsMoreThanOneMatchingElement(count)),
         };
@@ -609,7 +609,7 @@ public static class EnumerableExtensions
             }
         }
 
-        return Maybe<T>.None();
+        return Maybe<T>.None;
     }
 
     [StackTraceHidden]